+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## upcoming
+
+### Added
+- Added more descriptive error/help messages for required flags when --type is used in the 'ionosctl server create' command.
 ## [v6.9.6] – January 2026
 
 ### Added